Move files in preperation for mx-tester
