+++ test/harness/mjolnirTesting.ts
@@ -67,52 +67,10 @@ async function startSynpase() {
     console.info('starting synapse.')
     await compose.upOne('synapse_release', {config: composeConfig, log: true})
     await sleep(5000);
-    await registerUser('mjolnir', 'mjolnir', 'mjolnir', true);
     await registerUser('admin', 'admin', 'admin', true);
 }
 
-async function configureMjolnir() {
-    await fs.copyFile(path.join(__dirname, 'config', 'mjolnir', 'harness.yaml'), path.join(__dirname, '../../config/harness.yaml'));
-    await fs.rm(path.join(__dirname, 'mjolnir-data'), {recursive: true, force: true});
-    await fs.mkdir(path.join(__dirname, 'mjolnir-data')).catch (e => {
-        if (e.code === 'EEXIST') {
-            console.debug('mjolnir-data already exists');
-        } else {
-            throw e;
-        }
-    });
-    const pantalaimon = new PantalaimonClient(config.homeserverUrl, new MemoryStorageProvider());
-    const client = await pantalaimon.createClientWithCredentials(config.pantalaimon.username, config.pantalaimon.password);
-    await mjolnirSetup.ensureManagementRoomExists(client);
-    return await mjolnirSetup.ensureLobbyRoomExists(client);
-}
 
-async function startMjolnir() {
-    await configureMjolnir();
-    console.info('starting mjolnir');
-    await import('../../src/index');
-}
-
-async function registerUser(username: string, displayname: string, password: string, admin: boolean) {
-    let registerUrl = `${config.homeserverUrl}/_synapse/admin/v1/register`
-    let { data } = await axios.get(registerUrl);
-    let nonce = data.nonce!;
-    let mac = HmacSHA1(`${nonce}\0${username}\0${password}\0${admin ? 'admin' : 'notadmin'}`, 'REGISTRATION_SHARED_SECRET');
-    return await axios.post(registerUrl, {
-        nonce,
-        username,
-        displayname,
-        password,
-        admin,
-        mac: mac.toString()
-    }).catch(e => {
-        if (e.isAxiosError && e.response.data.errcode === 'M_USER_IN_USE') {
-            console.log(`${username} already registered, skipping`)
-        } else {
-            throw e;
-        }
-    });
-}
 
 export async function upHarness() {
     try {
